pyproject.toml
[project]
name = "news-collector"
version = "0.1.0"
description = "Agentic AI news collector using Claude and DuckDuckGo"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "ddgs>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
news-reader = "news_reader.agent:main"
news-writer = "news_reader.file_writer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/news_reader"]
